EWPP-2630: Refactor search response in mocks.
